requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
